src/main/java/code/kliangh/xml/utils/XmlReader.java
package code.kliangh.xml.utils;

import org.w3c.dom.Document;
import org.xml.sax.SAXException;

import javax.xml.parsers.DocumentBuilder;
import javax.xml.parsers.DocumentBuilderFactory;
import javax.xml.parsers.ParserConfigurationException;
import java.io.IOException;

/**
 * Created by kliangh on 28/06/2017.
 */
public class XmlReader {

    public Document readXML(String xmlPath) throws ParserConfigurationException, SAXException, IOException {

        Document document;
        DocumentBuilderFactory documentBuilderFactory = DocumentBuilderFactory.newInstance();

        DocumentBuilder documentBuilder = documentBuilderFactory.newDocumentBuilder();
        document = documentBuilder.parse(xmlPath);

        return document;
    }
}
